OAK-202 Simplify the code when possible
git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1364616 13f79535-47bb-0310-9956-ffa450edef68

+++ oak-mk/src/main/java/org/apache/jackrabbit/mk/blobs/AbstractBlobStore.java
@@ -331,7 +331,7 @@ public abstract class AbstractBlobStore implements Closeable, BlobStore, Cache.B
                 return;
             } else if (type == TYPE_DATA) {
                 int len = IOUtils.readVarInt(idStream);
-                IOUtils.skipFully(idStream, (int) len);
+                IOUtils.skipFully(idStream, len);
             } else if (type == TYPE_HASH) {
                 int level = IOUtils.readVarInt(idStream);
                 // totalLength
